Moved labels code to labels.py. Added some labels functions to utils.
