+++ vincentvanbot/params.py
@@ -7,3 +7,11 @@ JOBLIB_PATH_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__),'..','
 BUCKET_NAME='vincent-van-bot-bucket'
 BUCKET_JOBLIB_FOLDER='data'
 BUCKET_INITIAL_DATASET_FOLDER = 'data'
+
+# labels
+LABELS_SELECTION = [
+    'Illustration', 'History', 'Drawing', 'Mythology', 'Flower','Stock photography', 'Holy places',
+    'Event', 'Artifact', 'Ancient history', 'Wood', 'Vintage clothing', 'Tourist attraction',
+    'Sculpture','Font', 'Tree', 'Sky', 'Carving', 'Landscape', 'Metal', 'Middle ages', 'Prophet',
+    'Arch', 'Cloud', 'Building'
+    ]
